requirements.txt
pika==1.2.0
Flask==2.0.1
azure-mgmt-compute
azure-identity
